+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Handle external ID management [#364](https://github.com/rokwire/core-building-block/issues/364)
 
 ## [1.21.1] - 2022-03-17
 ### Fixed